requirements.txt
certifi==2019.9.11
Pillow==8.1.2
pipenv==2018.11.26
tqdm==4.59.0
virtualenv==16.7.7
virtualenv-clone==0.5.3
numpy~=1.20.1